Python/Gemini/Text/pyproject.toml
[project]
name = "text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.2.0",
    "python-dotenv>=1.0.1",
]
